requirements.txt
PyJWT>=2.10.1
pytest>=8.3.4
cryptography>=39.0.0